+++ src/app/beer_garden/requests.py
@@ -16,6 +16,7 @@ import time
 from asyncio import Future
 from builtins import str
 from copy import deepcopy
+from datetime import datetime
 from typing import Dict, List, Sequence, Union
 
 import six
@@ -667,7 +668,7 @@ def process_request(
         # Save after validation since validate can modify the request
         request = create_request(request)
 
-        logger.info(f"Publishing {request!r}")
+        logger.debug(f"Publishing {request!r}")
 
         if wait_event:
             request_map[request.id] = wait_event
@@ -876,6 +877,18 @@ def update_request(request: Request):
     return db.create(request)
 
 
+@publish_event(Events.REQUEST_UPDATED)
+def modify_request(request: Request = None, **kwargs):
+
+    # Clean commands are not run for modify, so have to add this in here instead
+    status_key = f"{request.status}_{config.get('garden.name')}"
+    if status_key not in request.metadata:
+        request.metadata[status_key] = int(datetime.utcnow().timestamp() * 1000)
+        kwargs["metadata"] = request.metadata
+
+    return db.modify(request, **kwargs)
+
+
 def process_wait(request: Request, timeout: float) -> Request:
     """Helper to process a request and wait for completion using a threading.Event
 
@@ -954,7 +967,27 @@ def handle_event(event):
         Events.REQUEST_CANCELED.name,
     ):
         # Only care about downstream garden
-        existing_request = db.query_unique(Request, id=event.payload.id)
+        requests = db.query(
+            Request,
+            filter_params={"id": event.payload.id},
+            include_fields=[
+                "id",
+                # Required to check if change in fields from child
+                "status",
+                "status_updated_at",
+                "target_garden",
+                "updated_at",
+                # Required for latency tracking
+                "metadata",
+                # Required for TEMP check
+                "command_type",
+            ],
+        )
+
+        if requests:
+            existing_request = requests[0]
+        else:
+            existing_request = None
 
         if existing_request:
             # Skip status that revert
@@ -970,30 +1003,43 @@ def handle_event(event):
             if existing_request is None:
                 # Attempt to create the request, if it already exists then continue on
                 try:
-                    db.create(event.payload)
+                    existing_request = db.create(event.payload)
                 except NotUniqueException:
                     pass
             elif event.name != Events.REQUEST_CREATED.name:
-                request_changed = False
+                request_changed = {}
                 # When we send child requests to child gardens where the parent was on
                 # the local garden we remove the parent before sending them. Only setting
                 # the subset of fields that change "corrects" the parent
                 for field in (
                     "status",
-                    "output",
-                    "error_class",
                     "status_updated_at",
                     "target_garden",
+                    "updated_at",
                 ):
                     new_value = getattr(event.payload, field)
 
                     if getattr(existing_request, field) != new_value:
-                        request_changed = True
-                        setattr(existing_request, field, new_value)
+                        request_changed[field] = new_value
+
+                # Add output fields only if the status changes to a compelted state
+                if "status" in request_changed:
+                    if event.payload.status in (
+                        "CANCELED",
+                        "SUCCESS",
+                        "ERROR",
+                        "INVALID",
+                    ):
+                        if event.payload.output:
+                            request_changed["output"] = event.payload.output
+                        if event.payload.error_class:
+                            request_changed["error_class"] = event.payload.error_class
 
                 if request_changed:
                     try:
-                        update_request(existing_request, _publish_error=False)
+                        existing_request = modify_request(
+                            existing_request, _publish_error=False, **request_changed
+                        )
                     except RequestStatusTransitionError:
                         pass
 
@@ -1020,7 +1066,7 @@ def clean_command_type_temp(request: Request, is_remote: bool):
             # Give Threading based requests a chance to pull the Request before deleting it
             time.sleep(0.5)
         db.delete(request)
-        return
+        return None
 
     # Delete any children that are TEMP once the current request is completed
     request.children = db.query(Request, filter_params={"parent": request})
@@ -1028,3 +1074,5 @@ def clean_command_type_temp(request: Request, is_remote: bool):
     for child in request.children:
         if child.command_type == "TEMP":
             db.delete(child)
+
+    return request