+++ src/app/beer_garden/requests.py
@@ -30,6 +30,7 @@ from brewtils.errors import (
 )
 from brewtils.models import Choices, Events, Operation, Request, RequestTemplate, System
 from brewtils.pika import PERSISTENT_DELIVERY_MODE
+from mongoengine import DoesNotExist
 from requests import Session
 
 import beer_garden.config as config
@@ -866,10 +867,14 @@ def invalid_request(request: Request = None):
 @publish_event(Events.REQUEST_UPDATED)
 def update_request(request: Request):
     if request.id:
-        request = db.update(request)
-    else:
-        request = db.create(request)
-    return request
+        try:
+            return db.update(request)
+        except DoesNotExist:
+            logger.warning(
+                f"Failed to update request {request.id}. Creating new request instead."
+            )
+
+    return db.create(request)
 
 
 @publish_event(Events.REQUEST_UPDATED)
@@ -984,7 +989,7 @@ def handle_event(event):
         else:
             existing_request = None
 
-        if existing_request:
+        if existing_request and existing_request.status != event.payload.status:
             # Skip status that revert
             if existing_request.status in ("CANCELED", "SUCCESS", "ERROR", "INVALID"):
                 return