+++ src/app/beer_garden/local_plugins/auto_brew/__main__.py
@@ -8,14 +8,11 @@ def main():
     module_name = sys.argv[1]
     class_name = sys.argv[2]
 
-    passedArgs = []
     passedKwargs = {}
     if len(sys.argv) > 3:
         for arg in sys.argv[3:]:
             if "=" in arg:
                 passedKwargs[arg.split("=")[0]] = arg.split("=")[1]
-            else:
-                passedArgs.append(arg)
 
     module = importlib.import_module(module_name)
     my_class = getattr(module, class_name)
@@ -27,7 +24,7 @@ def main():
         version=passedKwargs.get("VERSION", None),
     )
 
-    my_class_initialized = my_class(*passedArgs)
+    my_class_initialized = my_class()
 
     plugin = Plugin(
         name=my_class_initialized._bg_name,