+++ CHANGELOG.md
@@ -6,6 +6,7 @@ TBD
 
 - Reuse existing rabbit admin queue by purging messages instead of creating new queue
 - Fixed bug where unresponsive Garden API was referencing a non existing configuration for default. New default is -1.
+- Fixed bug where locally connected plugin commands ran as TEMP did not auto delete after completion
 
 # 3.25.0