+++ CHANGELOG.md
@@ -8,6 +8,7 @@ TBD
 - Reuse existing rabbit admin queue by purging messages instead of creating new queue
 - Fixed bug where Auto brew plugins did not parse NAME/VERSION properly from the beer.conf
 - Fixed bug where unresponsive Garden API was referencing a non existing configuration for default. New default is -1.
+- Fixed bug where locally connected plugin commands ran as TEMP did not auto delete after completion
 
 # 3.25.0