+++ CHANGELOG.md
@@ -7,6 +7,7 @@ TBD
 - Added support for cross-server parent/child requests
 - Reuse existing rabbit admin queue by purging messages instead of creating new queue
 - Fixed bug where unresponsive Garden API was referencing a non existing configuration for default. New default is -1.
+- Fixed bug where Garden unresponsvie metadata was cleared during child garden shutdown
 
 # 3.25.0