+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 TBD
 
+- Added auto refresh checkbox to requests page
 - Added support for cross-server parent/child requests
 - Reuse existing rabbit admin queue by purging messages instead of creating new queue
 - Fixed bug where Auto brew plugins did not parse NAME/VERSION properly from the beer.conf