+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 # 3.25.1
 
-TBD
+5/3/2024
 
 - Added auto refresh checkbox to requests page
 - Added support for cross-server parent/child requests