+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Beer Garden Changelog
 
+# 3.25.2
+
+TBD
+
+- Refactored Mongo pruner to use APScheduler
+
 # 3.25.1
 
 5/3/2024
@@ -7,7 +13,6 @@
 - Added auto refresh checkbox to requests page
 - Added support for cross-server parent/child requests
 - Reuse existing rabbit admin queue by purging messages instead of creating new queue
-- Refactored Mongo pruner to use APScheduler
 - Fixed bug where Auto brew plugins did not parse NAME/VERSION properly from the beer.conf
 - Fixed bug where unresponsive Garden API was referencing a non existing configuration for default. New default is -1.
 - Fixed bug where Garden unresponsive metadata was cleared during child garden shutdown